js: Split into smaller files
