+++ javascript/src/api/op_webhook_endpoint.ts
@@ -1,0 +1,75 @@
+import {
+    Configuration,
+    Ordering,
+    WebhookEndpointApi,
+    OperationalWebhookEndpointIn,
+    OperationalWebhookEndpointOut,
+    OperationalWebhookEndpointSecretIn,
+    OperationalWebhookEndpointSecretOut,
+    OperationalWebhookEndpointUpdate,
+    ListResponseOperationalWebhookEndpointOut,
+} from "../openapi";
+import { ListOptions, PostOptions } from "../util";
+
+export interface OperationalWebhookEndpointListOptions extends ListOptions {
+    order?: Ordering;
+}
+
+export class OperationalWebhookEndpoint {
+    private readonly api: WebhookEndpointApi;
+
+    public constructor(config: Configuration) {
+        this.api = new WebhookEndpointApi(config);
+    }
+
+    public list(
+        options?: OperationalWebhookEndpointListOptions
+    ): Promise<ListResponseOperationalWebhookEndpointOut> {
+        const iterator = options?.iterator ?? undefined;
+        return this.api.listOperationalWebhookEndpoints({ ...options, iterator });
+    }
+
+    public create(
+        endpointIn: OperationalWebhookEndpointIn,
+        options?: PostOptions
+    ): Promise<OperationalWebhookEndpointOut> {
+        return this.api.createOperationalWebhookEndpoint({
+            operationalWebhookEndpointIn: endpointIn,
+            ...options,
+        });
+    }
+
+    public get(endpointId: string): Promise<OperationalWebhookEndpointOut> {
+        return this.api.getOperationalWebhookEndpoint({ endpointId });
+    }
+
+    public update(
+        endpointId: string,
+        endpointUpdate: OperationalWebhookEndpointUpdate
+    ): Promise<OperationalWebhookEndpointOut> {
+        return this.api.updateOperationalWebhookEndpoint({
+            endpointId,
+            operationalWebhookEndpointUpdate: endpointUpdate,
+        });
+    }
+
+    public delete(endpointId: string): Promise<void> {
+        return this.api.deleteOperationalWebhookEndpoint({ endpointId });
+    }
+
+    public getSecret(endpointId: string): Promise<OperationalWebhookEndpointSecretOut> {
+        return this.api.getOperationalWebhookEndpointSecret({ endpointId });
+    }
+
+    public rotateSecret(
+        endpointId: string,
+        endpointSecretIn: OperationalWebhookEndpointSecretIn,
+        options?: PostOptions
+    ): Promise<void> {
+        return this.api.rotateOperationalWebhookEndpointSecret({
+            endpointId,
+            operationalWebhookEndpointSecretIn: endpointSecretIn,
+            ...options,
+        });
+    }
+}
