+++ javascript/src/util.ts
@@ -1,0 +1,8 @@
+export interface PostOptions {
+    idempotencyKey?: string;
+}
+
+export interface ListOptions {
+    iterator?: string | null;
+    limit?: number;
+}
